Move scripts into source tree
We currently have a couple of scripts that are called from cron that
depend on the clojars jar. This moves them into the source tree so they
can be updated with the source itself. The deploy script on the server
will copy them into the path of the clojars user on each deploy.